==> external_code/1.1_snomed_distance.ipynb <==
# Libraries
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)
import networkx as nx

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# -----------------------------
# Import snomed files
# -----------------------------

path = r'data/International/Relationship.txt'
international_relationship = pd.read_csv(path, delimiter="\t")

path = r'data/International/Description.txt'
international_description = pd.read_csv(path, delimiter="\t")

path = r'data/International/Concept.txt'
international_concept = pd.read_csv(path, delimiter="\t")

# -----------------------------
# Import files from icht
# -----------------------------
path = r'data/snomed_diagnosis_codes_morethan5.csv'
snomed_diagnosis_codes = pd.read_csv(path)

path = r'data/snomed_problem_codes_morethan5.csv'
snomed_problem_codes = pd.read_csv(path)

path = r'data/snomed_riskfactor_codes_morethan5.csv'
snomed_riskfactor_codes = pd.read_csv(path)

snomed_problem_codes

# Get list of codes
diagnosis_code_list = snomed_diagnosis_codes.DIAGNOSIS_CODE_SNOMED.values.tolist()
problem_code_list = snomed_problem_codes.PROBLEM.values.tolist()
riskfactor_code_list = snomed_riskfactor_codes.SNOMED_CODE.values.tolist()
# Combine
code_list = list(set(diagnosis_code_list) | set(problem_code_list) | set(riskfactor_code_list))

# Filter international snomed for disorders
description2 = international_description[international_description['term'].str.contains("\(disorder\)")==True]
# Get list of conceptIDs
conceptIDs = description2.conceptId.values.tolist()
# Filter relationship by disorder conceptIDs
relationship2 = international_relationship[(international_relationship['sourceId'].isin(conceptIDs)) | (international_relationship['destinationId'].isin(conceptIDs))]
# Filter relationship by 'is a' relationships
relationship2 = relationship2[relationship2['typeId'] == 116680003]
# Generate df for working out distances using codes
relationship2_2 = relationship2.copy()
relationship2_2.drop_duplicates(subset=['active', 'moduleId', 'sourceId', 'destinationId', \
                                      'relationshipGroup', 'typeId', 'characteristicTypeId', 'modifierId'], inplace=True)
relationship2_2 = relationship2_2[relationship2_2['active']==1]
# Remove codes from list that will not be in the graph 
sourceId_code_list = relationship2_2.sourceId.values.tolist()
destinationId_code_list = relationship2_2.destinationId.values.tolist()
Id_code_list = list(set(sourceId_code_list) | set(destinationId_code_list))
final_code_list = list(set(code_list) & set(Id_code_list))
len(Id_code_list)
len(code_list)
len(final_code_list)

# Graph with codes as nodes
id_disorder_graph = nx.from_pandas_edgelist(relationship2_2, 'sourceId', 'destinationId') # Graph with only current active relationships - can verify these paths online

# Work out path lengths and put in matrix
def path_lengths(graph, index_list):
    print('working...')
    np_path_matrix = np.zeros((len(index_list), len(index_list)))
    for a, b in enumerate(index_list):
        for c, d in enumerate(index_list):
            try:  
                path_length = nx.shortest_path_length(graph, source=b, target=d)
                np_path_matrix[a, c] = path_length
            except:
                np_path_matrix[a, c] = np.nan
    print('Done!')
    return np_path_matrix

id_disorder_np_path_matrix = path_lengths(id_disorder_graph, final_code_list)

# Create df of path lengths
path_df = pd.DataFrame(id_disorder_np_path_matrix, columns=final_code_list)
path_df.index = path_df.columns

# Do 1 / path_matrix to get numbers between 0 and 1\n
path_df2 = 1 / path_df
path_df2.replace([np.inf, -np.inf], 1, inplace=True)

# International 
id_disorder_np_path_matrix.shape

# Save - International
#path_df.to_csv('international_snomed_distance.csv')
#path_df2.to_csv('international_reciprocal_snomed_distance.csv')

# Adjustment
path = r'data/international_snomed_distance.csv'
international_snomed_distance = pd.read_csv(path, index_col=0)

international_snomed_distance

# Add one so path step of 1 does not become multiplication by one later
new_international_snomed_distance = international_snomed_distance + 1

new_international_snomed_distance

# Do 1 / path_matrix to get numbers between 0 and 1
new_international_reciprocal_snomed_distance = 1 / new_international_snomed_distance

# Save - International
#new_international_snomed_distance.to_csv('new_international_snomed_distance.csv')
#new_international_reciprocal_snomed_distance.to_csv('new_international_reciprocal_snomed_distance.csv')
