==> external_code/2.2_charlson.ipynb <==
# Libraries
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

## Import

# Import UMLS CUI to snomed map
iter_csv = pd.read_csv(r"data/MRCONSO.RRF", sep="|" , iterator=True, chunksize=10000000)
mrconso_df = pd.concat([chunk for chunk in iter_csv])
mrconso_df.columns = ['cui_code', 'language', 'term_status', 'LUI', 'string_type', 'SUI', 'atom_status', 'AUI', 'SAUI', 'SCUI', 'SDUI', 'SAB', 'TTY', 'code', 'STR', 'SRL', 'suppress', 'CVF', 'extra']
print('Done!')

# Import charlson index (downloaded online)
path = r'data/charlson_index_2.csv'
charlson_index = pd.read_csv(path)

charlson_index

charlson_index.info()

# Get list of conditions
comorbid_condition_list = charlson_index['Comorbid Condition'].unique().tolist()
# Remove nan
comorbid_condition_list = comorbid_condition_list[:-1]

# Turn off 'A value is trying to be set on a copy of a slice from a DataFrame'
pd.options.mode.chained_assignment = None  # default='warn'

# Convert to snomed code and combine with charlson index
snomed_charlson_index = pd.DataFrame()

for condition in comorbid_condition_list:
    print(condition)
    # Get icd codes
    temp_charlson_df = charlson_index[charlson_index['Comorbid Condition'] == condition]
    icd9_list = temp_charlson_df['ICD-9-CMDiagnosisCodes'].unique().tolist()
    icd10_list = temp_charlson_df['ICD-10-CADiagnosisCodes'].unique().tolist()
    icd_list = icd9_list + icd10_list
    # Filter for icd cdoes to find cui_code
    temp_mrconso_df = mrconso_df[(mrconso_df['SAB'].str.contains('icd', case=False)) & (mrconso_df['code'].isin(icd_list)) & (mrconso_df['language'] == 'ENG')]
    cui_code_list = temp_mrconso_df['cui_code'].unique().tolist()
    # Filter for cui_code to find snomed code
    temp_mrconso_df_2 = mrconso_df[(mrconso_df['cui_code'].isin(cui_code_list)) & (mrconso_df['language'] == 'ENG') & (mrconso_df['SAB'] == 'SNOMEDCT_US')]
    # Get somed codes
    temp_snomed_charlson_index = temp_mrconso_df_2[['code', 'STR']]
    temp_snomed_charlson_index.drop_duplicates(inplace=True)
    temp_snomed_charlson_index['comorbid_condition'] = condition
    temp_snomed_charlson_index['weight'] = temp_charlson_df['Weight'].iloc[0]
    snomed_charlson_index = pd.concat([snomed_charlson_index, temp_snomed_charlson_index], ignore_index=True)
print('Done!')    

snomed_charlson_index

# Import snomed_distance
path = r'data/new_international_snomed_distance.csv'
snomed_distance = pd.read_csv(path, index_col=0)

# Filter for those of interest to this research
icht_code_list = snomed_distance.columns.to_list()
filtered_snomed_charlson_index = snomed_charlson_index[snomed_charlson_index['code'].isin(icht_code_list)]

filtered_snomed_charlson_index

# Save
#filtered_snomed_charlson_index.to_csv('filtered_snomed_charlson_index.csv', index=False)
