==> external_code/4.3_metamap.ipynb <==
# Libraries
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

## Import

# Import indications
path = r'data/bnf_drug_indications_and_prescribing_medications.csv'
drug_indication_df = pd.read_csv(path)

# Import codes
path = r'data/snomed_diagnosis_codes_morethan5.csv'
snomed_diagnosis_codes = pd.read_csv(path)

path = r'data/snomed_problem_codes_morethan5.csv'
snomed_problem_codes = pd.read_csv(path)

path = r'data/snomed_riskfactor_codes_morethan5.csv'
snomed_riskfactor_codes = pd.read_csv(path)

# Import UMLS CUI to snomed map
iter_csv = pd.read_csv(r"data/MRCONSO.RRF", sep="|" , iterator=True, chunksize=10000000)
mrconso_df = pd.concat([chunk for chunk in iter_csv])
mrconso_df.columns = ['cui_code', 'language', 'term_status', 'LUI', 'string_type', 'SUI', 'atom_status', 'AUI', 'SAUI', 'SCUI', 'SDUI', 'SAB', 'TTY', 'snomed_code', 'STR', 'SRL', 'suppress', 'CVF', 'extra']
print('Done!')

# Import metamap output
metamap_output_df = pd.read_fwf('data/metamap_batch_output')
metamap_output_df.columns = ['metamap_text', 'other']

## Map UMLS CUI to SNOMED

mrconso_df

mrconso_df = mrconso_df[(mrconso_df['language'] == 'ENG') & (mrconso_df['SAB'] == 'SNOMEDCT_US')]
mrconso_df.reset_index(inplace=True, drop=True)
mrconso_df = mrconso_df[['cui_code', 'snomed_code']]
mrconso_df.drop_duplicates(inplace=True)

mrconso_df

metamap_df

# Import metamap output
# read_table works better than read_fwf
metamap_output_df = pd.read_table('data/metamap_batch_output')
metamap_output_df.columns = ['metamap_text']

# Filter for rows and text of interest
filtered_metamap_output_df = metamap_output_df[metamap_output_df['metamap_text'].str.contains('Processing', case=False) | metamap_output_df['metamap_text'].str.contains('Disease or Syndrome', case=False)]
# Fill na
filtered_metamap_output_df.fillna('', inplace=True)
# Reset index
filtered_metamap_output_df.reset_index(inplace=True, drop=True)

filtered_metamap_output_df

# First filter procesing lines 
processing_list = filtered_metamap_output_df[filtered_metamap_output_df['metamap_text'].str.contains('Processing', case=False)].index.to_list()

for x in processing_list:
    indication_text = filtered_metamap_output_df.iloc[x]['metamap_text'].split(":",1)[1]
    indication_text = indication_text.replace('"', '') 
    filtered_metamap_output_df.loc[x, 'indication'] = indication_text


# Second filter codes / disorders 
code_list = filtered_metamap_output_df[filtered_metamap_output_df['metamap_text'].str.contains('Disease or Syndrome', case=False)].index.to_list()

for x in code_list:
    start = 'C'
    end = ':'
    s  = filtered_metamap_output_df.iloc[x]['metamap_text']
    snomed_code_text = s[s.find(start)+len(start)-1:s.rfind(end)]
    snomed_code_text = snomed_code_text.replace(' ', '') 
    filtered_metamap_output_df.loc[x, 'cui_code'] = snomed_code_text
    
    disease_text = filtered_metamap_output_df.iloc[x]['metamap_text'].split(":",1)[1]
    filtered_metamap_output_df.loc[x, 'disease_or_syndrome'] = disease_text

# Forward fill indication column
filtered_metamap_output_df['indication'] = filtered_metamap_output_df['indication'].ffill()
# Drop rows where no snomed code
filtered_metamap_output_df = filtered_metamap_output_df[filtered_metamap_output_df['cui_code'].notna()]
# Drop full text column
filtered_metamap_output_df = filtered_metamap_output_df[['indication', 'cui_code', 'disease_or_syndrome']]
# Drop duplicates 
filtered_metamap_output_df.drop_duplicates(inplace=True)
# Reset index 
filtered_metamap_output_df.reset_index(inplace=True, drop=True)

filtered_metamap_output_df

indication_code_df

indication_code_df = pd.merge(filtered_metamap_output_df, mrconso_df)

indication_code_df

drug_indication_code_df

# Strip both so merge works properly 
drug_indication_df2 = drug_indication_df.copy()
indication_code_df2 = indication_code_df.copy()

# Remove text after BMI so obesity works properly
for x in range(len(drug_indication_df2)):
    indication_text = drug_indication_df2.iloc[x]['indication'].split("BMI",1)[0]
    drug_indication_df2.loc[x, 'indication'] = indication_text

for x in range(len(indication_code_df2)):
    indication_text = indication_code_df2.iloc[x]['indication'].split("BMI",1)[0]
    indication_code_df2.loc[x, 'indication'] = indication_text

drug_indication_df2['indication'] = drug_indication_df2['indication'].str.strip()
indication_code_df2['indication'] = indication_code_df2['indication'].str.strip()

drug_indication_code_df = pd.merge(drug_indication_df2, indication_code_df2)

# Rename and drop duplicates
drug_indication_code_df.drop_duplicates(inplace=True)
drug_indication_code_df.rename(columns={'drug_name':'bnf_drug_name', 'indication':'bnf_indication', 'prescribing_medication':'icht_prescribing_medication', 'disease_or_syndrome':'cui_disease_or_syndrome'}, inplace=True)

drug_indication_code_df

# See how many overlap 

snomed_problem_codes.rename(columns={'PROBLEM':'snomed_code'}, inplace=True)
snomed_diagnosis_codes.rename(columns={'DIAGNOSIS_CODE_SNOMED':'snomed_code'}, inplace=True)
snomed_riskfactor_codes.rename(columns={'SNOMED_CODE':'snomed_code'}, inplace=True)

snomed_problem_codes['snomed_code'] = snomed_problem_codes['snomed_code'].astype(str)
snomed_diagnosis_codes['snomed_code'] = snomed_diagnosis_codes['snomed_code'].astype(str)
snomed_riskfactor_codes['snomed_code'] = snomed_riskfactor_codes['snomed_code'].astype(str)

new_problem_df = pd.merge(drug_indication_code_df, snomed_problem_codes)
new_riskfactor_df = pd.merge(drug_indication_code_df, snomed_riskfactor_codes)
new_diagnosis_df = pd.merge(drug_indication_code_df, snomed_diagnosis_codes)

new_problem_df.snomed_code.nunique()
snomed_problem_codes.snomed_code.nunique()
new_problem_df.snomed_code.nunique()/snomed_problem_codes.snomed_code.nunique()
new_riskfactor_df.snomed_code.nunique()
snomed_riskfactor_codes.snomed_code.nunique()
new_riskfactor_df.snomed_code.nunique()/snomed_riskfactor_codes.snomed_code.nunique()
new_diagnosis_df.snomed_code.nunique()
snomed_diagnosis_codes.snomed_code.nunique()
new_diagnosis_df.snomed_code.nunique()/snomed_diagnosis_codes.snomed_code.nunique()

Fuzzy indication merge 

import difflib 
drug_indication_df2 = drug_indication_df.copy()
drug_indication_df2['new_indication'] = drug_indication_df2['indication'].apply(lambda x: difflib.get_close_matches(x, indication_code_df['indication'], n=1, cutoff=0.8))#[0])

# select first element in list for fuzzy indication merge
for x in range(len(drug_indication_df2)):
    if len(drug_indication_df2.iloc[x]['new_indication']) >= 1:
        indication_2_text = drug_indication_df2.iloc[x]['new_indication'][0]
        drug_indication_df2.loc[x, 'indication_2'] = indication_2_text
    else:
        indication_2_text = drug_indication_df2.iloc[x]['indication']
        drug_indication_df2.loc[x, 'indication_2'] = indication_2_text


drug_indication_df2

# New merge with indication_2
drug_indication_code_df_2 = pd.merge(drug_indication_df2[['drug_name', 'prescribing_medication', 'indication_2']], indication_code_df, how='inner', left_on='indication_2', right_on='indication')

# Rename and drop duplicates
drug_indication_code_df_2.drop_duplicates(inplace=True)
drug_indication_code_df_2.rename(columns={'drug_name':'bnf_drug_name', 'indication':'bnf_indication', 'prescribing_medication':'icht_prescribing_medication', 'disease_or_syndrome':'cui_disease_or_syndrome'}, inplace=True)
drug_indication_code_df_2.drop('indication_2', axis=1, inplace=True)

drug_indication_code_df_2

# Merge fuzzy with normal

drug_indication_code_df_3 = drug_indication_code_df_2.copy()
drug_indication_code_df_3['bnf_indication'] = drug_indication_code_df_3['bnf_indication'].str.strip()
drug_indication_code_df_3 = pd.concat([drug_indication_code_df, drug_indication_code_df_3])

drug_indication_code_df_3.drop_duplicates(inplace=True)
drug_indication_code_df_3.reset_index(drop=True, inplace=True)

drug_indication_code_df_3

# See how many overlap

new_problem_df_2 = pd.merge(drug_indication_code_df_3, snomed_problem_codes)
new_riskfactor_df_2 = pd.merge(drug_indication_code_df_3, snomed_riskfactor_codes)
new_diagnosis_df_2 = pd.merge(drug_indication_code_df_3, snomed_diagnosis_codes)

new_problem_df_2.snomed_code.nunique()
snomed_problem_codes.snomed_code.nunique()
new_problem_df_2.snomed_code.nunique()/snomed_problem_codes.snomed_code.nunique()
new_riskfactor_df_2.snomed_code.nunique()
snomed_riskfactor_codes.snomed_code.nunique()
new_riskfactor_df_2.snomed_code.nunique()/snomed_riskfactor_codes.snomed_code.nunique()
new_diagnosis_df_2.snomed_code.nunique()
snomed_diagnosis_codes.snomed_code.nunique()
new_diagnosis_df_2.snomed_code.nunique()/snomed_diagnosis_codes.snomed_code.nunique()

# Check numbers match 

new_problem_df_4 = pd.merge(indication_code_df, snomed_problem_codes)
new_riskfactor_df_4 = pd.merge(indication_code_df, snomed_riskfactor_codes)
new_diagnosis_df_4 = pd.merge(indication_code_df, snomed_diagnosis_codes)

new_problem_df_4.snomed_code.nunique()
snomed_problem_codes.snomed_code.nunique()
new_problem_df_4.snomed_code.nunique()/snomed_problem_codes.snomed_code.nunique()
new_riskfactor_df_4.snomed_code.nunique()
snomed_riskfactor_codes.snomed_code.nunique()
new_riskfactor_df_4.snomed_code.nunique()/snomed_riskfactor_codes.snomed_code.nunique()
new_diagnosis_df_4.snomed_code.nunique()
snomed_diagnosis_codes.snomed_code.nunique()
new_diagnosis_df_4.snomed_code.nunique()/snomed_diagnosis_codes.snomed_code.nunique()

Save

# Save
#drug_indication_code_df.to_csv('drugs_indications_and_codes.csv', index=False)

# Save
#drug_indication_code_df_3.to_csv('drugs_indications_and_codes_fuzzy.csv', index=False)
